changed lookup to be one function
